+++ Model/ConfigProvider.php
@@ -150,9 +150,6 @@ class ConfigProvider extends CcGenericConfigProvider
 //                    'checkoutLogoUrl'           => $checkout_logo,
                     'isTestMode'                => $this->moduleConfig->isTestModeEnabled(),
                     'countryId'                 => $this->moduleConfig->getQuoteCountryCode(),
-                    'updateQuotePM'             => $this->urlBuilder
-                        ->getUrl('nuvei_checkout/payment/UpdateQuotePaymentMethod'),
-                    
                     'isPaymentPlan'             => !empty($payment_plan_data) ? 1 : 0,
                     'useDevSdk'                 => $this->moduleConfig->getConfigValue('use_dev_sdk'),
                     
@@ -183,8 +180,6 @@ class ConfigProvider extends CcGenericConfigProvider
                         'maskCvv'                   => true,
                         'i18n'                      => $this->moduleConfig->getCheckoutTransl(),
                         'blockCards'                => $blocked_cards,
-//                        'billingAddress'            => $billing_address,
-//                        'userData'                  => ['billingAddress' => $billing_address],
                     ],
                 ],
             ],